refactor: #673 rag/main.py を router/service/store 境界で分割する
FastAPI エントリを薄くし、models・services・routers へ責務を移す。
HTTP 契約とプロンプト意味は不変。テスト互換のため main 再エクスポートを維持する。
