+++ rag/routers/__init__.py
@@ -1,0 +1,1 @@
+"""RAG FastAPI ルーター。"""
