+++ rag/routers/resume.py
@@ -1,0 +1,117 @@
+"""履歴書レビューエンドポイント。"""
+from __future__ import annotations
+
+import json
+import logging
+import os
+from typing import Generator
+
+from fastapi import APIRouter
+from fastapi.responses import StreamingResponse
+from openai import OpenAI
+
+from models import ReviewRequest, ReviewResponse
+from services.sanitize import _sanitize_company_name_for_query, _sanitize_job_title
+
+logger = logging.getLogger("main")
+
+router = APIRouter()
+
+
+@router.post("/resume/review/stream")
+def review_resume_stream(request: ReviewRequest) -> StreamingResponse:
+    """RAGレポートをSSEでストリーミング配信するエンドポイント。"""
+    import main as m
+
+    safe_company_for_prompt = _sanitize_company_name_for_query(request.company_name)
+    role_label = _sanitize_job_title(request.job_title) if request.job_title else "指定なし"
+
+    # コンテキスト収集（キャッシュ/DeepResearch/Web Search）
+    retrieved, context_source = m._gather_context(request)
+
+    source_labels = {
+        "deep_research": "OpenAI Deep Research（o3-deep-research）",
+        "web_search": "OpenAI Web Search（gpt-4o-search-preview）",
+        "cache": "ベクトルDBキャッシュ（Chroma / 以前の検索結果）",
+        "none": "事前学習データのみ（外部検索なし）",
+    }
+    source_label = source_labels.get(context_source, context_source)
+
+    def generate() -> Generator[str, None, None]:
+        api_key = os.getenv("OPENAI_API_KEY")
+        if not api_key:
+            yield "data: {}\n\n".format(
+                json.dumps({"type": "error", "message": "OPENAI_API_KEY not set"}, ensure_ascii=False))
+            return
+
+        model = os.getenv("OPENAI_REVIEW_MODEL", "gpt-4o-mini")
+        context_block = "\n\n".join(retrieved) if retrieved else "（外部情報なし）"
+
+        prompt = (
+            "以下の企業の採用観点に照らして、候補者の履歴書のレビューレポートを日本語で作成してください。\n\n"
+            "【企業名】{company}\n"
+            "【職種】{role}\n"
+            "【企業情報（参考）】\n{context}\n\n"
+            "【履歴書テキスト】\n{resume}\n\n"
+            "以下のフォーマットに従って出力してください:\n\n"
+            "【企業別レビュー報告書】\n---\n"
+            "#### ■ 対象企業\n{company}\n\n"
+            "#### ■ この企業が求めている核心的要素\n- ...\n\n"
+            "#### ■ 履歴書の最適化アドバイス\n"
+            "- **強みの再定義**: ...\n"
+            "- **不足している情報の補足**: ...\n\n"
+            "#### ■ 職種別アドバイス（{role}）\n"
+            "この職種特有の評価ポイントを3点以上、具体的に記述してください。\n\n"
+            "#### ■ 修正後の自己PRイメージ\n...\n\n"
+            "#### ■ 情報の信頼度・参照元\n"
+            "- 情報ソース: {source}\n"
+            "- 注意: 外部情報に基づく内容は変化する可能性があります。最新情報は企業公式サイトで確認してください。\n"
+        ).format(
+            company=safe_company_for_prompt,
+            role=role_label,
+            context=context_block,
+            resume=request.resume_text[:m.RESUME_REVIEW_INPUT_CHAR_LIMIT],
+            source=source_label,
+        )
+
+        try:
+            client = OpenAI(api_key=api_key)
+            stream = client.chat.completions.create(
+                model=model,
+                messages=[
+                    {"role": "system", "content": "あなたはプロのキャリアアドバイザーです。"},
+                    {"role": "user", "content": prompt},
+                ],
+                stream=True,
+                max_tokens=1500,
+                temperature=0.3,
+            )
+            for chunk in stream:
+                delta = chunk.choices[0].delta
+                if delta.content:
+                    yield "data: {}\n\n".format(
+                        json.dumps({"type": "chunk", "text": delta.content}, ensure_ascii=False))
+            yield "data: {}\n\n".format(json.dumps({"type": "done"}, ensure_ascii=False))
+        except Exception as exc:
+            logger.error("review_stream generate error: %s", exc)
+            yield "data: {}\n\n".format(json.dumps({"type": "error", "message": str(exc)}, ensure_ascii=False))
+
+    return StreamingResponse(generate(), media_type="text/event-stream")
+
+
+@router.post("/resume/review", response_model=ReviewResponse)
+def review_resume(request: ReviewRequest) -> ReviewResponse:
+    import main as m
+
+    role_label = request.job_title or "指定なし"
+    retrieved, context_source = m._gather_context(request)
+
+    report = m.run_crewai(
+        resume_text=request.resume_text,
+        company_name=request.company_name,
+        job_title=role_label,
+        context_docs=retrieved,
+        context_source=context_source,
+    )
+
+    return ReviewResponse(report=report)
