+++ rag/services/__init__.py
@@ -1,0 +1,1 @@
+"""RAG サービス層（埋め込み・調査・CrewAI・ヒント・ES 等）。"""
