+++ rag/services/es_review.py
@@ -1,0 +1,87 @@
+"""ES（エントリーシート）添削。"""
+from __future__ import annotations
+
+import logging
+import os
+from typing import List
+
+from fastapi import HTTPException
+
+from models import ESReviewResponse
+
+logger = logging.getLogger("main")
+
+
+def _run_es_review(
+        es_text: str,
+        question_type: str,
+        company_name: str,
+        context_docs: List[str],
+) -> ESReviewResponse:
+    import json as _json
+    import main as m
+
+    api_key = os.getenv("OPENAI_API_KEY")
+    if not api_key:
+        raise HTTPException(status_code=500, detail="OPENAI_API_KEY is required")
+    client = m.OpenAI(api_key=api_key)
+    model = os.getenv("OPENAI_CHAT_MODEL", "gpt-4o")
+    has_company = bool(company_name.strip())
+    context_text = "\n\n".join(context_docs) if context_docs else ""
+    company_section = (
+        f"\n\n【企業情報】\n{context_text[:2000]}" if context_text else ""
+    )
+    company_fit_key = (
+        '"company_fit_score": <1-10の整数: 企業の価値観・求める人物像との適合度>'
+        if has_company
+        else '"company_fit_score": null'
+    )
+    system_prompt = (
+        "あなたは就職活動の専門アドバイザーです。"
+        "学生のES文章を添削し、以下のJSONのみを返してください。説明文は不要です。"
+    )
+    user_prompt = (
+            f"【質問種別】{question_type}\n"
+            f"【ES文章】\n{es_text}"
+            + (f"\n\n【志望企業】{company_name}" if has_company else "")
+            + company_section
+            + f"""
+
+以下のJSONフォーマットで添削結果を返してください:
+{{
+  "specificity_score": <1-10の整数: 具体的な数値・エピソード・固有名詞が含まれているか>,
+  "star_score": <1-10の整数: Situation/Task/Action/Resultの構造が揃っているか>,
+  {company_fit_key},
+  "length_balance_score": <1-10の整数: 文字数・各要素のバランスが適切か>,
+  "feedback": "<具体性・STAR準拠・企業適合性・文字数について400字程度でアドバイス。企業名が指定されている場合は企業特化アドバイスを含めてください>",
+  "improved_text": "<元の文章を改善したバージョン（元の文字数の110〜130%を目安）>",
+  "company_strategy": "<企業特化の対策アドバイス（企業名なしは null。指定時は約200〜400字）>"
+}}"""
+    )
+    try:
+        resp = client.chat.completions.create(
+            model=model,
+            messages=[
+                {"role": "system", "content": system_prompt},
+                {"role": "user", "content": user_prompt},
+            ],
+            temperature=0.3,
+            max_tokens=1200,
+            response_format={"type": "json_object"},
+        )
+        data = _json.loads(resp.choices[0].message.content or "{}")
+        company_fit = data.get("company_fit_score")
+        if company_fit is not None:
+            company_fit = max(1, min(10, int(company_fit)))
+        return ESReviewResponse(
+            specificity_score=max(1, min(10, int(data.get("specificity_score", 5)))),
+            star_score=max(1, min(10, int(data.get("star_score", 5)))),
+            company_fit_score=company_fit,
+            length_balance_score=max(1, min(10, int(data.get("length_balance_score", 5)))),
+            feedback=str(data.get("feedback", "")),
+            improved_text=str(data.get("improved_text", "")),
+            company_strategy=(str(data.get("company_strategy")) if data.get("company_strategy") is not None else None),
+        )
+    except Exception as exc:
+        logger.warning("es review failed error=%s", exc)
+        raise HTTPException(status_code=500, detail=f"ES review failed: {exc}")
